api/requirements.txt
fastapi
pydantic
psycopg2
python-dotenv
uvicorn
